refactor(integrations): keep sanitized-inputs.ts in place, only add code-scanning
Reverts the previous split structure to a more conservative approach:
keep sanitized-inputs.ts (and its check id 'sanitized_inputs') and just
strip the code-scanning logic out of it. Code scanning becomes a single
new self-contained file (code-scanning.ts).

This preserves continuity for existing customers:
- IntegrationCheckRun records under 'sanitized_inputs' continue to be
  the same automation
- IntegrationConnection.metadata.disabledTaskChecks entries that
  reference 'sanitized_inputs' remain valid

Removes the helper modules (input-validation.ts, validation-libraries.ts,
code-scanning-detector.ts) introduced in the previous commit.
