+++ packages/integration-platform/src/manifests/github/checks/sanitized-inputs.ts
@@ -1,23 +1,55 @@
 /**
- * Input Validation Check
+ * Sanitized Inputs Check
  *
- * Verifies repositories use a modern validation/sanitization library.
+ * Ensures repositories use a modern validation/sanitization library.
  * Supports monorepos by scanning all package.json / requirements.txt /
- * pyproject.toml / composer.json files in the repository tree.
+ * pyproject.toml / composer.json files.
  *
- * Detection covers JS/TS, Python, and PHP ecosystems.
+ * Validation-library detection covers JS/TS, Python, and PHP ecosystems.
  */
 
 import { TASK_TEMPLATES } from '../../../task-mappings';
 import type { IntegrationCheck } from '../../../types';
 import type { GitHubRepo, GitHubTreeEntry, GitHubTreeResponse } from '../types';
 import { parseRepoBranch, targetReposVariable } from '../variables';
-import {
-  JS_VALIDATION_PACKAGES,
-  PHP_VALIDATION_PACKAGES,
-  PY_VALIDATION_PACKAGES,
-  VALIDATION_TARGET_FILES,
-} from './validation-libraries';
+
+const JS_VALIDATION_PACKAGES = [
+  'zod',
+  'yup',
+  'joi',
+  '@effect/schema',
+  'effect',
+  'valibot',
+  'ajv',
+  'class-validator',
+  'io-ts',
+  'superstruct',
+  'runtypes',
+];
+
+const PY_VALIDATION_PACKAGES = [
+  'pydantic',
+  'marshmallow',
+  'cerberus',
+  'voluptuous',
+  'jsonschema',
+  'schematics',
+  'typeguard',
+];
+
+const PHP_VALIDATION_PACKAGES = [
+  'laravel/framework',
+  'respect/validation',
+  'symfony/validator',
+  'vlucas/valitron',
+];
+
+const TARGET_FILES = [
+  'package.json',
+  'requirements.txt',
+  'pyproject.toml',
+  'composer.json',
+];
 
 interface GitHubFileResponse {
   content: string;
@@ -43,88 +75,9 @@ const getFileName = (path: string): string => {
   return parts[parts.length - 1] ?? path;
 };
 
-const escapeRegex = (s: string): string => s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-
-const checkPackageJson = (content: string, filePath: string): ValidationMatch | null => {
-  try {
-    const pkg = JSON.parse(content);
-    const deps = {
-      ...(pkg.dependencies || {}),
-      ...(pkg.devDependencies || {}),
-    };
-    for (const candidate of JS_VALIDATION_PACKAGES) {
-      if (deps[candidate]) {
-        return { library: candidate, file: filePath };
-      }
-    }
-  } catch {
-    // Invalid JSON, skip
-  }
-  return null;
-};
-
-const checkPythonFile = (
-  content: string,
-  filePath: string,
-  fileName: string,
-): ValidationMatch | null => {
-  // Parse line-by-line so we can strip comments and match package names as
-  // standalone tokens (e.g. don't match "schema" inside "jsonschema" or inside
-  // a prose comment).
-  //
-  // Comment syntax differs between the two Python dependency file formats:
-  //   - pyproject.toml (TOML): `#` starts a comment anywhere outside a string.
-  //     Dependency values are always quoted, so stripping at any `#` is safe
-  //     — the package name in `"name @ url#egg=name"` appears before the URL.
-  //   - requirements.txt (pip): `#` only starts a comment when preceded by
-  //     whitespace (or at line start). This preserves VCS URL fragments like
-  //     `git+https://...#egg=pydantic` used for editable installs.
-  const isTOML = fileName === 'pyproject.toml';
-  const stripCommentRegex = isTOML ? /#.*$/ : /(^|\s)#.*$/;
-  const lines = content.split('\n');
-  for (const rawLine of lines) {
-    const line = rawLine.replace(stripCommentRegex, '').toLowerCase();
-    if (!line.trim()) continue;
-    for (const candidate of PY_VALIDATION_PACKAGES) {
-      const escaped = escapeRegex(candidate.toLowerCase());
-      // Match the package name as a standalone token. Leading context: start
-      // of line or a separator commonly preceding a package name (whitespace,
-      // quote, bracket, comma, semicolon, or `=` for `#egg=name` VCS syntax).
-      // Trailing context: end of line or a separator that can follow a
-      // package name (whitespace, version operator, bracket, quote, comma,
-      // semicolon).
-      const pattern = new RegExp(
-        `(?:^|[\\s"'\\[,;=])${escaped}(?:$|[\\s=<>!~\\[\\]"',;])`,
-      );
-      if (pattern.test(line)) {
-        return { library: candidate, file: filePath };
-      }
-    }
-  }
-  return null;
-};
-
-const checkComposerJson = (content: string, filePath: string): ValidationMatch | null => {
-  try {
-    const pkg = JSON.parse(content);
-    const deps = {
-      ...(pkg.require || {}),
-      ...(pkg['require-dev'] || {}),
-    };
-    for (const candidate of PHP_VALIDATION_PACKAGES) {
-      if (deps[candidate]) {
-        return { library: candidate, file: filePath };
-      }
-    }
-  } catch {
-    // Invalid JSON, skip
-  }
-  return null;
-};
-
-export const inputValidationCheck: IntegrationCheck = {
-  id: 'input_validation',
-  name: 'Input Validation',
+export const sanitizedInputsCheck: IntegrationCheck = {
+  id: 'sanitized_inputs',
+  name: 'Sanitized Inputs',
   description:
     'Verifies repositories use a supported input-validation library (JS/TS, Python, or PHP). Scans entire repository including monorepo subdirectories.',
   service: 'code-security',
@@ -141,7 +94,7 @@ export const inputValidationCheck: IntegrationCheck = {
       ctx.fail({
         title: 'No repositories selected',
         description:
-          'Select at least one repository to monitor in the integration settings so we can verify input validation.',
+          'Select at least one repository to monitor in the integration settings so we can verify sanitized inputs.',
         resourceType: 'integration',
         resourceId: 'github',
         severity: 'low',
@@ -183,14 +136,89 @@ export const inputValidationCheck: IntegrationCheck = {
       }
     };
 
+    const checkPackageJson = (content: string, filePath: string): ValidationMatch | null => {
+      try {
+        const pkg = JSON.parse(content);
+        const deps = {
+          ...(pkg.dependencies || {}),
+          ...(pkg.devDependencies || {}),
+        };
+        for (const candidate of JS_VALIDATION_PACKAGES) {
+          if (deps[candidate]) {
+            return { library: candidate, file: filePath };
+          }
+        }
+      } catch {
+        // Invalid JSON, skip
+      }
+      return null;
+    };
+
+    const escapeRegex = (s: string): string =>
+      s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+
+    const checkPythonFile = (
+      content: string,
+      filePath: string,
+      fileName: string,
+    ): ValidationMatch | null => {
+      // Parse line-by-line; strip comments and match package names as standalone
+      // tokens to avoid matching "schema" inside "jsonschema" or in prose comments.
+      // requirements.txt only treats `#` as a comment when whitespace-preceded —
+      // that preserves VCS URL fragments like `git+https://...#egg=pydantic`.
+      // pyproject.toml uses normal `#` since dependency values are always quoted.
+      const isTOML = fileName === 'pyproject.toml';
+      const stripCommentRegex = isTOML ? /#.*$/ : /(^|\s)#.*$/;
+      const lines = content.split('\n');
+      for (const rawLine of lines) {
+        const line = rawLine.replace(stripCommentRegex, '').toLowerCase();
+        if (!line.trim()) continue;
+        for (const candidate of PY_VALIDATION_PACKAGES) {
+          const escaped = escapeRegex(candidate.toLowerCase());
+          // Match the package name as a standalone token. Leading context: start
+          // of line or a separator commonly preceding a package name (whitespace,
+          // quote, bracket, comma, semicolon, or `=` for `#egg=name` VCS syntax).
+          // Trailing context: end of line or a separator that can follow a
+          // package name (whitespace, version operator, bracket, quote, comma,
+          // semicolon).
+          const pattern = new RegExp(
+            `(?:^|[\\s"'\\[,;=])${escaped}(?:$|[\\s=<>!~\\[\\]"',;])`,
+          );
+          if (pattern.test(line)) {
+            return { library: candidate, file: filePath };
+          }
+        }
+      }
+      return null;
+    };
+
+    const checkComposerJson = (content: string, filePath: string): ValidationMatch | null => {
+      try {
+        const pkg = JSON.parse(content);
+        const deps = {
+          ...(pkg.require || {}),
+          ...(pkg['require-dev'] || {}),
+        };
+        for (const candidate of PHP_VALIDATION_PACKAGES) {
+          if (deps[candidate]) {
+            return { library: candidate, file: filePath };
+          }
+        }
+      } catch {
+        // Invalid JSON, skip
+      }
+      return null;
+    };
+
     const findValidationLibraries = async (
       repoName: string,
       tree: GitHubTreeEntry[],
     ): Promise<ValidationMatch[]> => {
       const matches: ValidationMatch[] = [];
 
+      // Find all target files in the tree
       const targetEntries = tree.filter(
-        (entry) => entry.type === 'blob' && VALIDATION_TARGET_FILES.includes(getFileName(entry.path)),
+        (entry) => entry.type === 'blob' && TARGET_FILES.includes(getFileName(entry.path)),
       );
 
       for (const entry of targetEntries) {
@@ -239,7 +267,7 @@ export const inputValidationCheck: IntegrationCheck = {
         });
       } else {
         const checkedFiles = tree
-          .filter((e) => e.type === 'blob' && VALIDATION_TARGET_FILES.includes(getFileName(e.path)))
+          .filter((e) => e.type === 'blob' && TARGET_FILES.includes(getFileName(e.path)))
           .map((e) => e.path);
 
         const supportedLibraries = [
